wip: moved Penalties struct up one module and set up a TODO for Noah
